+++ library/trans.c
@@ -261,7 +261,7 @@ void dnet_check_tree(struct dnet_node *n, int kill)
 		dnet_trans_remove_nolock(&n->trans_root, t);
 
 		dnet_log(n, DNET_LOG_ERROR, "%s: freeing transaction %llu: err: %d.\n",
-				dnet_dump_id(t->cmd.id), t->trans, err);
+				dnet_dump_id(t->cmd.id), (unsigned long long)t->trans, err);
 
 		t->cmd.status = err;
 		t->cmd.size = 0;
